Update key.rs
Refactored to utilize std::io::Cursor for reading through the byte array read in from the KEY file.

Updated Key::fromFile to create a Cursor and call Key::fromCursor.
Implemented fromCursor for Key, BifEntry, and ResourceEntry.
Updated FromFileTest to verify the existence, and minimally validate the contents, of BIF entries and Resource entries.