Update bytes.rs
Added function readToReal to omit any garbage data when reading bytes which are meant to be parsed as a string.

Renamed readString to parseString.

Added readBytesExact to make reading via cursor.read_exact less verbose.

Implemented a new readString to be a one-step process, utilizing readBytesExact, readToReal, and parseString.

Updated readResRef and readName to rely on readString.