Update are.rs
Added Are::readWed() method for loading this area's WED file; Until this method is called, the wed member will be None; Accepts a reference to a ResourceManager, since presumably an instance will already be available to have loaded the ARE file in the first place.

Updated test to load the WED file and verify that it is Some.