Update tis.rs
Updated to utilize Colors struct to translate from BGRA to RGBA

Modified PaletteSize constant to be 256 instead of 1024, since now the colors are being read as a list of u32 values rather than the entire set of bytes all at once.

Store list of Colors alongside the palette list; Probably don't need the palette list anymore but might as well keep it around for now as the original form of the data.

Update toBytes to iterate over self.colors not self.palette.

Implement ReadIntoSelf for Tis.
Added function new to simplify generating a new default instance with a tileCount value.

Updated how tileset entries are read in Bif to account for these changes.