Update Cargo.toml
Added global lints to Cargo.toml.
Removed lint assertions from the files.
Added 'static lifetime to constant string slices where appropriate.
Removed mut operator where it is unnecessary.

Updated FreeDimensions to not call drop, as it does nothing to a Copy type, as well as mark the parameter as not being used; I'll need to look into the best way to deal with this sort of a situation but for now I'll maintain this method to reinforce the pattern of needing to send the request a resource be freed when it is retrieved from across the FFI barrier.